add cpu

+++ srcs/func.c
@@ -6,7 +6,7 @@
 /*   By: hcaspar <hcaspar@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/28 10:57:25 by hcaspar           #+#    #+#             */
-/*   Updated: 2017/03/30 17:40:39 by hcaspar          ###   ########.fr       */
+/*   Updated: 2017/03/30 19:33:46 by hcaspar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 void				zoom_in(int x, int y, t_env *e)
 {
-	float			xscale;
-	float			yscale;
-	float			ztemp;
-	float			xtemp;
-	float			ytemp;
+	double			xscale;
+	double			yscale;
+	double			ztemp;
+	double			xtemp;
+	double			ytemp;
 
 	ztemp = e->v.z;
 	xtemp = e->v.x;
@@ -37,11 +37,11 @@ void				zoom_in(int x, int y, t_env *e)
 
 void				zoom_out(int x, int y, t_env *e)
 {
-	float			xscale;
-	float			yscale;
-	float			ztemp;
-	float			xtemp;
-	float			ytemp;
+	double			xscale;
+	double			yscale;
+	double			ztemp;
+	double			xtemp;
+	double			ytemp;
 
 	ztemp = e->v.z;
 	xtemp = e->v.x;
